+++ cmp/chart/index.ts
@@ -1,2 +1,3 @@
 export * from './Chart';
 export * from './ChartModel';
+export * from './Types';