+++ mobile/cmp/button/index.ts
@@ -13,6 +13,7 @@ export * from './grid/ExpandToLevelButton';
 export * from './FeedbackButton';
 export * from './ThemeToggleButton';
 export * from './LogoutButton';
+export * from './OptionsButton';
 export * from './RefreshButton';
 export * from './RestoreDefaultsButton';
 export * from './NavigatorBackButton';